refactor(frontend): extract flightDateOf/prevUtcDay to a leaf module, break the useFlightTrack↔flightFilter import cycle
Task 7's OR-filter fix had useFlightTrack.ts import prevUtcDay from
flightFilter.ts, which already imported flightDateOf from useFlightTrack.ts —
a two-file cycle safe only via function-declaration hoisting. Both helpers
now live in flightDates.ts (a leaf module importing nothing from FlightTracker
siblings); every consumer (flightFilter.ts, useFlightTrack.ts, useRouteIndex.ts,
useAltitudeProfile.ts, and their tests) imports from there instead. No
re-exports kept — flightDateOf/prevUtcDay tests moved into flightDates.test.ts.
