+++ packages/frontend/src/Applications/FlightTracker/flightFilter.ts
@@ -1,5 +1,5 @@
 import type { FlightPosition } from "../../Providers/MediaStream/MediaStreamContext";
-import { flightDateOf } from "./useFlightTrack";
+import { flightDateOf, prevUtcDay } from "./flightDates";
 
 // Filter criteria for the Filter Flights window (issue #188). Values come from
 // dropdowns fed by live data, so matching is exact; "" means "any". Criteria
@@ -44,21 +44,6 @@ export type RouteIndex = Map<string, RouteIndexRow>;
 
 export function routeKey(flight: string, flightDate: string): string {
 	return `${flight}|${flightDate}`;
-}
-
-// flight_tracks.flight_date is the BTS *local* departure date, but a
-// FlightPosition's start_date (and flightDateOf, which takes its UTC date
-// component) is a UTC instant. An evening flight departing e.g. 8:30 PM ET on
-// 9/12 has flight_date = "2001-09-12" while every one of its samples is dated
-// "2001-09-13" UTC (00:30Z onward) — so joining strictly on
-// routeKey(flight, flightDateOf(start_date)) misses it. Across every US
-// timezone in this dataset, flight_date is always either the sample's UTC
-// date or the day before it, so a lookup falls back one UTC day before
-// giving up.
-export function prevUtcDay(flightDate: string): string {
-	return new Date(Date.parse(`${flightDate}T00:00:00Z`) - 86_400_000)
-		.toISOString()
-		.slice(0, 10);
 }
 
 // Joins a streamed position to its route-index row across that local/UTC
